NIFI-7819 - Moved StateProviderContext to nifi-framework-api.
