+++ nifi-framework-api/src/main/java/org/apache/nifi/components/state/annotation/StateProviderContext.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.nifi.controller.state.providers.zookeeper;
+package org.apache.nifi.components.state.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
